+++ CHANGELOG.md
@@ -2,6 +2,26 @@
 
 **New features:**
 
+- A new `.pipe()` method on Decoder allows you to pass the output of one decoder as input
+  to another:
+  ```tsx
+  string
+    .transform((s) => s.split(',')) // transform first...
+    .pipe(array(nonEmptyString)); //   ...then validate that result
+  ```
+  This was previously possible already with `.then`, but it was hard to work with.
+- The new `.pipe()` can also dynamically select another decoder, based on the input:
+  ```tsx
+  string
+    .transform((s) => s.split(',').map(Number)) // transform first...
+    .pipe((tup) =>
+      tup.length === 2
+        ? point2d
+        : tup.length === 3
+          ? point3d
+          : never('Invalid coordinate'),
+    );
+  ```
 - The `formatShort` formatter will now quote error positions with single quotes, which
   makes them more human-readable in JSON responses.